Phase 5a — Split oversized internal/tui files per CONVENTIONS.md §9
Pure, mechanical file reorganization with zero behavior change. Splits the
oversized monolithic components into the `_model`/`_update`/`_view` layout and
collapses the forbidden `xxxpanel.go`/`xxxpanelmodel.go` pairs into the same
three-file split:

- dashboard.go     -> dashboard_{model,update,view}.go
- planeditor.go    -> planeditor_{model,update,view}.go
- reviewpanel.go + reviewpanelmodel.go     -> reviewpanel_{model,update,view}.go
- shippingpanel.go + shippingpanelmodel.go -> shippingpanel_{model,update,view}.go

Code moved verbatim by the §9 boundary (model = structs/constructor/state
builders; update = Update + handlers + owned Msg types; view = View + render
helpers). A line-multiset equivalence check confirms each component's union of
new files is identical to its old file(s). No logic, signatures, or behavior
changed; full race suite, vet, lint, and e2e all pass unchanged.

Establishes the target layout for the follow-up state-relocation (5b) and
App.Update thinning (5c) PRs.
