+++ internal/tui/reviewpanel_view.go
@@ -840,3 +840,28 @@ func wrapText(s string, maxWidth int) []string {
 	}
 	return lines
 }
+
+// View renders the review panel — either the spec overlay or the main panel.
+func (m *reviewPanelModel) View(svc PanelServices) string {
+	if m == nil || m.session == nil {
+		return ""
+	}
+	if m.specOverlay {
+		plan, _ := m.session.CachedPlan()
+		return renderReviewSpecOverlay(m.session, plan, m.specOverlayScroll, m.width, m.height)
+	}
+	entry := svc.ReviewCache(m.repoPath, m.session.ID)
+	prDraftInFlight := svc.prDraftInFlightFor(m.session.ID, m.repoPath)
+	var checkState *checksTabState
+	if svc.ValidationRuns != nil {
+		if run := svc.ValidationRuns(m.session.ID); run != nil {
+			checkState = &checksTabState{
+				checks:  run.checks,
+				results: run.results,
+				cursor:  m.checksCursor,
+				scroll:  m.checksScroll,
+			}
+		}
+	}
+	return renderReviewPanel(m.session, entry, m.width, m.height, m.taskCursor, prDraftInFlight, m.activeTab, checkState, m.now)
+}
